fixed a bug

+++ roachgram/users/signals.py
@@ -27,7 +27,8 @@ def removeDir(path):
 # when user account deleted all of its uploads gets deleted
 @receiver(pre_delete , sender=User)
 def on_user_delete(sender , instance , **kwargs):
-    removeDir(f"uploads/{instance.username}")
+    if (instance.profile != "default-profile.png"):
+        removeDir(f"uploads/{instance.username}")
 
 
 def deleteFile(path):
